extracting a dedicated media decrypter
